Update _params.py
Unify help text vocabulary to always refer to parameters of the format 'a=b' as a (secure) (environment) variable assignment. Remove misleading mention of "space-separated".

`--environment-variables 'a=hello b=c'`

is interpreted as 'set the environment variable with name "a" to the value "hello b=c"'. It is not equivalent to

`--environment-variables 'a=hello' 'b=c'`

The separation of the arguments is handled by the shell and might not involve the space character at all depending on the shell and the settings.